refactor: refactor place component:
