+++ src/components/hashtag-list.tsx
@@ -8,7 +8,12 @@ interface HashtagListProps {
 
 const HashtagList = ({ hashtags, className }: HashtagListProps) => {
   return (
-    <div className={cn('flex items-center gap-2 overflow-x-scroll', className)}>
+    <div
+      className={cn(
+        'flex items-center gap-2 overflow-x-scroll no-scrollbar',
+        className,
+      )}
+    >
       {hashtags.map((hashtag) => (
         <Chip
           className="whitespace-nowrap"
